feature: hooks
Hooks are blocks of independent code that will be executed during certain parts of the application lifecycle; for instance, by registering a 'request' hook on a channel, you can place custom metadata on every outgoing request to set something like a trace ID. On the server, you can define a 'decode' hook on the router that allows you to manipulate a decode record _before_ it is passed to a service method for us.

The same types of hooks can be defined on both the client and server side; just think of them inversely. So where a 'decode' hook on the server lets you inspect and modify an incoming record from a client, on the client the hook it allows you to do the same for a record received from the server.

Each hook has access to the request/response context, but cannot modify things such as HTTP headers. They also have access to an environment object which can be used to pass around reusable components like database clients (this is the same one you provide to a router.)

You also have the ability to short-circuit hook execution by not calling 'next'